ShouldThrowValidationExceptionOnAddIfCompanyIsNullAndLogItAsync -> FAIL
